requirements_whole_project.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.0
anyio==3.6.2
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
Babel==2.11.0
beautifulsoup4==4.11.1
bertopic==0.15.0
bleach==5.0.1
blinker==1.5
bqplot==0.12.36
branca==0.6.0
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
cligj==0.7.2
colour==0.1.5
commonmark==0.9.1
Cython==0.29.36
dataclasses-json==0.5.7
defusedxml==0.7.1
dnspython==2.4.2
faiss-cpu==1.7.4
fastjsonschema==2.16.2
filelock==3.8.0
Fiona==1.8.13.post1
folium==0.13.0
fonttools==4.25.0
frozenlist==1.3.3
fsspec==2023.10.0
GDAL==3.0.2
gdown==4.5.3
gensim==4.3.2
geojson==2.5.0
gitdb==4.0.10
GitPython==3.1.29
hdbscan==0.8.33
huggingface-hub==0.19.4
idna==3.4
importlib-metadata==5.0.0
ipyevents==2.0.1
ipyfilechooser==0.6.0
ipyleaflet==0.17.2
ipysheet==0.6.0
ipython-genutils==0.2.0
ipytree==0.2.2
ipywidgets==7.7.2
Jinja2==3.1.2
json5==0.9.10
jsonschema==4.17.1
keplergl==0.3.2
langchain==0.0.190
leafmap==0.13.2
llvmlite==0.41.1
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
mistune==2.0.4
mpmath==1.3.0
multidict==6.0.4
munkres==1.1.4
mypy-extensions==1.0.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.58.1
openai==0.27.2
openapi-schema-pydantic==1.2.4
pandas==1.5.1
pandocfilters==1.5.0
Pillow==9.2.0
pipx==1.0.0
plotly==5.11.0
prometheus-client==0.15.0
protobuf==3.20.3
pyarrow==10.0.1
pycparser==2.21
PyCRS==1.0.2
pydantic==1.10.8
pydeck==0.8.0
pymongo==4.5.0
Pympler==1.0.1
pynndescent==0.5.11
pyrsistent==0.19.2
pyshp==2.3.1
PySocks==1.7.1
pystac==1.6.1
pystac-client==0.5.1
python-box==6.1.0
python-dotenv==1.0.0
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.3.22
requests==2.28.1
rich==12.6.0
safetensors==0.4.0
scipy==1.9.3
scooby==0.7.0
semver==2.13.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.15
streamlit==1.15.1
streamlit-chat==0.0.2.2
sympy==1.12
tenacity==8.1.0
terminado==0.17.0
tiktoken==0.3.2
tinycss2==1.2.1
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.64.1
traitlets==5.5.0
traittypes==0.2.1
transformers==4.35.2
typing-inspect==0.9.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
umap-learn==0.5.5
urllib3==1.26.13
userpath==1.8.0
validators==0.20.0
watchdog==2.1.9
webencodings==0.5.1
websocket-client==1.4.2
whitebox==2.2.0
whiteboxgui==2.2.0
widgetsnbextension==3.6.1
xyzservices==2022.9.0
yarl==1.8.2
zipp==3.10.0
huggingface_hub
